fix slack chat and add conversation

+++ backend/src/Modules/SlackChat/SlackChat/Workspaces/Features/DeleteChannel/DeleteChannelHandler.cs
@@ -11,14 +11,25 @@ public class DeleteChannelHandler
 {
   public async Task<DeleteChannelResult> Handle(DeleteChannelCommand command, CancellationToken cancellationToken)
   {
+    var userId = user.GetUserId();
     var workspace = await dbContext.Workspaces
       .Where(x => x.Id == command.WorkspaceId)
+      .Include(x => x.Members.Where(x => x.UserId == userId))
       .Include(x => x.Channels.Where(t => t.Id == command.ChannelId))
       .FirstOrDefaultAsync(cancellationToken)
       ?? throw new WorkspaceNotFoundException(command.WorkspaceId);
+    if (workspace.Members.Count == 0)
+    {
+      throw new BadRequestException("Unauthorized");
+    }
+    var member = workspace.Members[0];
+    if (member.Role != MemberRole.Admin)
+    {
+      throw new BadRequestException("Unauthorized");
+    }
 
     workspace.RemoveChannel(command.ChannelId);
-
+    await dbContext.SaveChangesAsync(cancellationToken);
     return new DeleteChannelResult(true);
   }
 }
